requirements.txt
aws-cdk-lib
aws-xray-sdk
boto3
constructs
invoke
mypy
pytest
pytest-cov
pytest-mock
ruff
simplelens
types-boto3
